move old cache implementation to allensdk/api/warehouse_cache
so that there is no confusion whether or not the new and old cache
implementations should be able to work with each other
