+++ allensdk/api/queries/ontologies_api.py
@@ -34,7 +34,7 @@
 # POSSIBILITY OF SUCH DAMAGE.
 #
 from .rma_template import RmaTemplate
-from ..cache import cacheable
+from allensdk.api.warehouse_cache.cache import cacheable
 
 from allensdk.core.structure_tree import StructureTree
