+++ allensdk/brain_observatory/behavior/session_apis/data_io/behavior_lims_api.py
@@ -6,7 +6,7 @@ import pytz
 
 import pandas as pd
 
-from allensdk.api.cache import memoize
+from allensdk.api.warehouse_cache.cache import memoize
 from allensdk.brain_observatory.behavior.session_apis.abcs.\
     data_extractor_base.behavior_data_extractor_base import \
     BehaviorDataExtractorBase
