+++ allensdk/core/cell_types_cache.py
@@ -37,7 +37,7 @@ import os
 from six import string_types
 
 from allensdk.config.manifest_builder import ManifestBuilder
-from allensdk.api.cache import Cache, get_default_manifest_file
+from allensdk.api.warehouse_cache.cache import Cache, get_default_manifest_file
 from allensdk.api.queries.cell_types_api import CellTypesApi
 
 from . import json_utilities as json_utilities
