+++ allensdk/internal/api/queries/mouse_connectivity_api_prerelease.py
@@ -1,4 +1,4 @@
-from allensdk.api.cache import Cache, cacheable
+from allensdk.api.warehouse_cache.cache import Cache, cacheable
 from allensdk.api.queries.grid_data_api import GridDataApi
 from allensdk.api.queries.mouse_connectivity_api import MouseConnectivityApi
