+++ allensdk/test/api/test_cacheable.py
@@ -35,7 +35,7 @@
 #
 import pytest
 from mock import MagicMock, patch, mock_open
-from allensdk.api.cache import Cache, cacheable
+from allensdk.api.warehouse_cache.cache import Cache, cacheable
 from allensdk.api.queries.rma_api import RmaApi
 import pandas as pd
 from six.moves import builtins
